Fix/delete buttons and timesheet approval

+++ src/main/java/io/softwaregarage/hris/admin/views/PositionListView.java
@@ -17,25 +17,49 @@ import com.vaadin.flow.router.PageTitle;
 import com.vaadin.flow.router.Route;
 
 import io.softwaregarage.hris.admin.dtos.PositionDTO;
+import io.softwaregarage.hris.admin.dtos.UserDTO;
 import io.softwaregarage.hris.admin.services.PositionService;
+import io.softwaregarage.hris.admin.services.UserService;
 import io.softwaregarage.hris.commons.views.MainLayout;
+import io.softwaregarage.hris.profile.dtos.PositionProfileDTO;
+import io.softwaregarage.hris.profile.services.PositionProfileService;
+import io.softwaregarage.hris.utils.SecurityUtil;
 
 import jakarta.annotation.Resource;
 import jakarta.annotation.security.RolesAllowed;
 
 import org.vaadin.lineawesome.LineAwesomeIcon;
 
+import java.util.List;
+
 @RolesAllowed({"ROLE_ADMIN", "ROLE_HR_MANAGER", "ROLE_HR_SUPERVISOR"})
 @PageTitle("Positions")
 @Route(value = "position-list", layout = MainLayout.class)
 public class PositionListView extends VerticalLayout {
-    @Resource private final PositionService positionService;
+    @Resource
+    private final PositionService positionService;
+
+    @Resource
+    private final PositionProfileService positionProfileService;
+
+    @Resource
+    private final UserService userService;
+
+    private UserDTO userDTO;
 
     private TextField searchFilterTextField;
     private Grid<PositionDTO> positionDTOGrid;
 
-    public PositionListView(PositionService positionService) {
+    public PositionListView(PositionService positionService,
+                            PositionProfileService positionProfileService,
+                            UserService userService) {
         this.positionService = positionService;
+        this.positionProfileService = positionProfileService;
+        this.userService = userService;
+
+        if (SecurityUtil.getAuthenticatedUser() != null) {
+            userDTO = userService.getByUsername(SecurityUtil.getAuthenticatedUser().getUsername());
+        }
 
         this.add(buildHeaderToolbar(), buildPositionDTOGrid());
         this.setSizeFull();
@@ -113,42 +137,57 @@ public class PositionListView extends VerticalLayout {
         deleteButton.addThemeVariants(ButtonVariant.LUMO_PRIMARY, ButtonVariant.LUMO_ERROR);
         deleteButton.addClickListener(buttonClickEvent -> {
             if (positionDTOGrid.getSelectionModel().getFirstSelectedItem().isPresent()) {
-                // Show the confirmation dialog.
-                ConfirmDialog confirmDialog = new ConfirmDialog();
-                confirmDialog.setHeader("Delete Position");
-                confirmDialog.setText(new Html("""
+                PositionDTO selectedPositionDTO = positionDTOGrid.getSelectionModel().getFirstSelectedItem().get();
+
+                List<PositionProfileDTO> positionProfileDTOList = positionProfileService.findByParameter(selectedPositionDTO.getCode());
+                if (positionProfileDTOList.isEmpty()) {
+                    // Show the confirmation dialog.
+                    ConfirmDialog confirmDialog = new ConfirmDialog();
+                    confirmDialog.setHeader("Delete Position");
+                    confirmDialog.setText(new Html("""
                                                <p>
                                                WARNING! Before deleting this position, be sure that there are no
                                                employees assigned to it. Are you sure you want to delete the selected
                                                position?
                                                </p>
                                                """));
-                confirmDialog.setConfirmText("Yes, Delete it.");
-                confirmDialog.setConfirmButtonTheme("error primary");
-                confirmDialog.addConfirmListener(confirmEvent -> {
-                    // Get the selected position and delete it.
-                    PositionDTO selectedPositionDTO = positionDTOGrid.getSelectionModel().getFirstSelectedItem().get();
-                    positionService.delete(selectedPositionDTO);
+                    confirmDialog.setConfirmText("Yes, Delete it.");
+                    confirmDialog.setConfirmButtonTheme("error primary");
+                    confirmDialog.addConfirmListener(confirmEvent -> {
+                        // Get the selected position and delete it.
+                        positionService.delete(selectedPositionDTO);
 
-                    // Refresh the data grid from the backend after the delete operation.
-                    positionDTOGrid.getDataProvider().refreshAll();
+                        // Refresh the data grid from the backend after the delete operation.
+                        positionDTOGrid.getDataProvider().refreshAll();
 
+                        // Show notification message.
+                        Notification notification = Notification.show("You have successfully deleted the selected position.",
+                                5000,
+                                Notification.Position.TOP_CENTER);
+                        notification.addThemeVariants(NotificationVariant.LUMO_SUCCESS);
+
+                        // Close the confirmation dialog.
+                        confirmDialog.close();
+                    });
+                    confirmDialog.setCancelable(true);
+                    confirmDialog.setCancelText("No");
+                    confirmDialog.open();
+                } else {
                     // Show notification message.
-                    Notification notification = Notification.show("You have successfully deleted the selected position.",
-                                                          5000,
-                                                                  Notification.Position.TOP_CENTER);
-                    notification.addThemeVariants(NotificationVariant.LUMO_SUCCESS);
-
-                    // Close the confirmation dialog.
-                    confirmDialog.close();
-                });
-                confirmDialog.setCancelable(true);
-                confirmDialog.setCancelText("No");
-                confirmDialog.open();
+                    Notification notification = Notification.show("You cannot delete the selected position. There are employees assigned to it.",
+                            5000,
+                            Notification.Position.TOP_CENTER);
+                    notification.addThemeVariants(NotificationVariant.LUMO_ERROR);
+                }
             }
         });
 
-        rowToolbarLayout.add(viewButton, editButton, deleteButton);
+        if (userDTO.getRole().equals("ROLE_ADMIN")) {
+            rowToolbarLayout.add(viewButton, editButton, deleteButton);
+        } else {
+            rowToolbarLayout.add(viewButton, editButton);
+        }
+
         rowToolbarLayout.setJustifyContentMode(JustifyContentMode.CENTER);
         rowToolbarLayout.getStyle().set("flex-wrap", "wrap");
